Resolve a final Clippy warning
Here we needed to revise the code so that the initial evaluation of the
mutable borrow always comes after each use of the shared borrow that it
conflicts with.  In this case, this just meant that I needed to pull
the value out of the function call, so that the calculation of the
value passed into plugins.replace was computed _strictly_ before the
call began?

Something like that.

This is in relation to rust-lang/rust#59159, which is effectively going
to remove two-phase borrows in this context.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>
Tested-by: Kristofer Rye <kristofer.rye@gmail.com>